+++ src/Managers/Assistants/CreateAssistant.php
@@ -6,7 +6,7 @@ use Illuminate\Contracts\View\Factory;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\View;
 use Thinktomorrow\Chief\Forms\Fields\Validation\FieldValidator;
-use Thinktomorrow\Chief\Forms\Layouts\Layout;
+use Thinktomorrow\Chief\Forms\Layouts\PageLayout;
 use Thinktomorrow\Chief\ManagedModels\Events\ManagedModelCreated;
 use Thinktomorrow\Chief\Managers\Exceptions\NotAllowedManagerAction;
 use Thinktomorrow\Chief\Managers\Routes\ManagedRoute;
@@ -47,7 +47,7 @@ trait CreateAssistant
         View::share('model', $model);
         View::share('resource', $this->resource);
 
-        View::share('layout', Layout::make($this->resource->fields($model))->model($model));
+        View::share('layout', PageLayout::make($this->resource->fields($model))->model($model));
 
         return $this->resource->getCreatePageView();
     }
@@ -75,7 +75,7 @@ trait CreateAssistant
     {
         $model = $this->managedModelClassInstance(...$this->resource->getInstanceAttributes($request));
 
-        $fields = Layout::make($this->resource->fields($model))
+        $fields = PageLayout::make($this->resource->fields($model))
             ->model($model)
             ->getFields()
             ->filterByNotTagged(['edit', 'not-on-model-create', 'not-on-create']);