+++ src/Plugins/TimeTable/App/Http/DateController.php
@@ -7,7 +7,7 @@ use Thinktomorrow\Chief\App\Http\Controllers\Controller;
 use Thinktomorrow\Chief\Forms\App\Actions\SaveFields;
 use Thinktomorrow\Chief\Forms\App\Queries\Fields;
 use Thinktomorrow\Chief\Forms\Fields\Validation\FieldValidator;
-use Thinktomorrow\Chief\Forms\Layouts\Layout;
+use Thinktomorrow\Chief\Forms\Layouts\PageLayout;
 use Thinktomorrow\Chief\Plugins\TimeTable\Domain\Events\DateCreated;
 use Thinktomorrow\Chief\Plugins\TimeTable\Domain\Events\DateDeleted;
 use Thinktomorrow\Chief\Plugins\TimeTable\Domain\Events\DateUpdated;
@@ -33,7 +33,7 @@ class DateController extends Controller
         [$model, $fields] = $this->getModelAndFields();
 
         // Enrich fields with the layout components
-        $forms = Layout::make($model->fields($model))->model($model)->getComponents();
+        $forms = PageLayout::make($model->fields($model))->model($model)->getComponents();
         $fields = $forms[0]->getComponents();
 
         return view('chief-timetable::dates.create', [
@@ -82,7 +82,7 @@ class DateController extends Controller
         [$model, $fields] = $this->getModelAndFields($dateId);
 
         // Enrich fields with the layout components
-        $forms = Layout::make($model->fields($model))->model($model)->getComponents();
+        $forms = PageLayout::make($model->fields($model))->model($model)->getComponents();
         $fields = $forms[0]->getComponents();
 
         return view('chief-timetable::dates.edit', [